+++ index.d.ts
@@ -1,0 +1,3 @@
+declare module "bitly";
+
+export * from "./src/index";
