+++ src/index.ts
@@ -1,0 +1,9 @@
+export { BitlyClient } from "./bitly";
+export { doRequest, generateUrl, sortUrlsAndHash } from "./lib";
+export {
+  BitlyConfig,
+  BitlyUrlQueryParams,
+  BitlyResponseData,
+  BitlyResponse,
+  BitlyError
+} from "./types";
